requirements.txt
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
beautifulsoup4==4.12.2
bs4==0.0.1
charset-normalizer==3.3.2
frozenlist==1.4.1
multidict==6.0.4
PyYAML==6.0.1
requests==2.31.0
soupsieve==2.5
tqdm==4.66.1
yarl==1.9.4
chardet==5.2.0